requirements.txt
asgiref==3.7.2
cffi==1.15.1
cfgv==3.3.1
cryptography==41.0.2
distlib==0.3.7
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.2
gunicorn==21.2.0
identify==2.5.26
nodeenv==1.8.0
numpy==1.25.1
packaging==23.1
pandas==2.0.3
platformdirs==3.9.1
pre-commit==3.3.3
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
virtualenv==20.24.2
whitenoise==6.5.0
